tests/wpt/tests/css/css-tables/parsing/border-collapse-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Table Module Test: parsing border-collapse with invalid values</title>
<link rel="help" href="https://drafts.csswg.org/css-tables/#propdef-border-collapse">
<meta name="assert" content="border-collapse supports only the grammar 'separate | collapse'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("border-collapse", "none");
test_invalid_value("border-collapse", "separate collapse");
</script>
</body>
</html>
